LayoutTests/fast/shapes/shape-inside/shape-inside-polygon-layout.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/subpixel-utils.js"></script>
<style>
    #shape-inside {
        position: relative;
        width: 594px;
        height: 500px;
        -webkit-shape-inside: polygon(1.58% 1.42%, 99.42% 1.42%, 98.83% 33.33%, 50.67% 34.08%, 99.50% 66.62%, 98.83% 99.48%, 2.25% 99.07%);
        font: 14.5px/1 Ahem, sans-serif;
        color: green;
    }

    #shape-outline {
        position: absolute;
        display: block;
        top: 0px;
        left: 0px;
        width: 594px;
        height: 500px;
    }
</style>
</head>
<body>
<body>
    <div id="shape-inside">
        <svg id="shape-outline" viewbox="0 0 100 100" preserveAspectRatio="none" xmlns="http://www.w3.org/2000/svg">
            <polygon id="shape-polygon" points="1.58,1.42 99.42,1.42 98.83,33.33 50.67,34.08 99.50,66.62, 98.83,99.48 2.25,99.07" stroke="blue" vector-effect="non-scaling-stroke" stroke-width="1" fill="none"/>
        </svg>
        <span id="a">X</span><br/>
        <span id="b">X</span><br/>
        <span id="c">X</span><br/>
        <span id="d">X</span><br/>
    </div>
    <div id="console"></div>
</body>
<script>
  function shapeInsideRect(elementId) 
  {
      var s = document.getElementById("shape-inside").getBoundingClientRect();
      var r = document.getElementById(elementId).getBoundingClientRect();
      return {left: r.left - s.left, top: r.top - s.top, width: r.width, height: r.height};
  }

  // The top of the polygon is a horzontal edge at y=1.42%, and its parent's height is 500px.  The top of
  // the first line, span "a", will align with the top of the polygon. The ExclusionShape::firstIncludedIntervalLogicalTop()
  // methods use ceiledLayoutUnit() to ensure that their return value is within the shape.

  var lineATop = Math.floor(SubPixelLayout.ceilSnapToLayoutUnit(.0142 * 500)); 

  // Verify that all four one-character lines are vertically adjacent and 15 pixels high.
  var quiet = true; // PASS output depends on SubPixelLayout.isEnabled()
  shouldBe("shapeInsideRect('a').top", String(lineATop), quiet);
  shouldBe("shapeInsideRect('b').top", String(lineATop + 15), quiet);
  shouldBe("shapeInsideRect('c').top", String(lineATop + 30), quiet);
  shouldBe("shapeInsideRect('d').top", String(lineATop + 45), quiet);
</script>
</html>
